+++ src/pages/profile/index.tsx
@@ -2,8 +2,9 @@ import { HeaderLink, Layout } from "@/components/Layout";
 import { Card, InfiniteScroll, NoResults } from "@/components/List";
 import { HeaderSearchBar, SEARCH_QUERY_KEY, SortDialog, useParseSort } from "@/components/Search";
 import { Star, UserPic } from "@/components/UI";
-import { fetchNextPage, getQueryParam, minutesToMs } from "@/utils";
+import { fetchNextPage, minutesToMs } from "@/utils";
 import { api, type RouterInputs, type RouterOutputs } from "@/utils/api";
+import { getQueryParam } from "@/utils/query";
 import type { NextPageWithLayout } from "@/utils/type";
 import type { Session } from "next-auth";
 import { useSession } from "next-auth/react";
