+++ src/server/api/routers/product.ts
@@ -6,7 +6,8 @@ import { createTRPCRouter, protectedProcedure } from "@/server/api/trpc";
 import { cacheProductNames, getProductNames } from "@/server/redis";
 import { getFileUrl } from "@/server/uploadthing";
 import getScrapedProducts from "@/server/utils/scrapers";
-import { isUrl, mostCommonItems } from "@/utils";
+import { isUrl } from "@/utils";
+import { mostCommonItems } from "@/utils/array";
 import { and, asc, desc, eq, exists, gt, like, lt, or, sql, type SQL } from "drizzle-orm";
 import { z } from "zod";
 import { throwDefaultError } from "../utils";
