+++ src/utils/index.ts
@@ -1,52 +1,10 @@
 import type { UseTRPCInfiniteQueryResult } from "@trpc/react-query/shared";
-import type { NextRouter } from "next/router";
-import type { Query } from "nextjs-routes";
+import type { Option, Some } from "./type";
 
 export const browser = typeof window !== "undefined";
 
-export function getQueryParam(query: Query[string]) {
-  if (query === undefined) return undefined;
-  return Array.isArray(query) ? query.at(-1) : query;
-}
-
-type SetQueryParamOptions = { push?: boolean };
-/** Deletes falsy values */
-export function setQueryParam(
-  router: NextRouter,
-  key: string,
-  value?: string | null,
-  options?: SetQueryParamOptions,
-) {
-  if (!value) {
-    delete router.query[key];
-  } else {
-    router.query[key] = value;
-  }
-
-  if (options?.push) {
-    void router.push({ query: router.query });
-    return;
-  }
-  void router.replace({ query: router.query });
-}
-
 export function clamp(min: number, value: number, max: number) {
   return Math.min(max, Math.max(min, value));
-}
-
-export function nonEmptyArray<T>(array: T[]): array is [T, ...T[]] {
-  return !!array.length;
-}
-
-/** Checks if element is an array - narrows the type of checked element */
-export function includes<T>(array: readonly T[], element: unknown): element is T {
-  return array.includes(element as T);
-}
-
-export function indexOf(array: readonly unknown[], value: unknown) {
-  const index = array.indexOf(value);
-  if (index === -1) return null;
-  return index;
 }
 
 export function minutesToMs(minutes: number) {
@@ -64,23 +22,6 @@ export function isUrl(url: string) {
 
 export function isNonEmptyString(value: unknown): value is string {
   return !!value && typeof value === "string";
-}
-
-export function mostCommonItems(count: number) {
-  return function <T>(arr: T[]): NonNullable<T>[] {
-    const counter = new Map<NonNullable<T>, number>();
-    for (const element of arr) {
-      if (element == null) continue;
-
-      const count = counter.get(element) ?? 0;
-      counter.set(element, count + 1);
-    }
-
-    return Array.from(counter)
-      .sort(([, a], [, b]) => b - a)
-      .slice(0, count)
-      .map(([v]) => v);
-  };
 }
 
 export function fetchNextPage<Q extends UseTRPCInfiniteQueryResult<unknown, unknown>>({
@@ -105,4 +46,8 @@ export function setIntersection<T>(a: Set<T>, b: Set<T>) {
 export function isSetEqual<T>(a: Set<T>, b: Set<T>) {
   const intersection = setIntersection(a, b);
   return intersection.size === a.size && intersection.size === b.size;
+}
+
+export function isSome<O extends Option<unknown>>(option: O): option is Extract<O, Some<unknown>> {
+  return option.ok;
 }
